runner: move cmd processing thread work back to runner
The aio patches moved the cmd processing thread work from runner
to libtcmu. This moves it back since it is tcmu-runner specific.
